Refactor serialization of Cypher results
