+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/ErrorMessagesTest.scala
@@ -22,10 +22,10 @@ package org.neo4j.cypher
 import org.apache.commons.lang3.SystemUtils
 import org.hamcrest.CoreMatchers._
 import org.junit.Assert._
-import org.neo4j.cypher.internal.compiler.v3_0.commands.expressions.StringHelper
+import org.neo4j.cypher.internal.compiler.v3_0.CypherSerializer
 import org.neo4j.cypher.internal.frontend.v3_0.helpers.StringHelper
 
-class ErrorMessagesTest extends ExecutionEngineFunSuite with StringHelper {
+class ErrorMessagesTest extends ExecutionEngineFunSuite with CypherSerializer {
 
   test("fails on incorrect unicode literal") {
     expectSyntaxError(
